Fixing memory problems (Arduino)

+++ DartsBoard/src/SimpleMap.h
@@ -5,23 +5,23 @@
 
 template<typename T, typename U>
 class SimpleMap {
-	int currentSize = 0;
-	int maxSize;
+	uint8_t currentSize = 0;
+	uint8_t maxSize;
 	
 	T* key;
 	U* value;
 	
 public:
-	SimpleMap(int maxSize) {
+	SimpleMap(uint8_t maxSize) {
 		this->maxSize = maxSize;
-	
+
 		key = new T[maxSize];
 		value = new U[maxSize];
 	}
 	
 	bool insert(T e1, U e2) {
 		if (currentSize < maxSize) {
-			key[currentSize] = e1;	
+			key[currentSize] = e1;
 			value[currentSize] = e2;		
 		
 			currentSize++;
@@ -32,7 +32,7 @@ public:
 	}
 	
 	U getValueByKey(T key) {
-		for (int i = 0; i < currentSize; i++) {
+		for (uint8_t i = 0; i < currentSize; i++) {
 			if (key == this->key[i] ) {
 				return this->value[i];
 			}
@@ -42,7 +42,7 @@ public:
 	}
 
 	T getKeyByValue(U value) {
-		for (int i = 0; i < currentSize; i++) {
+		for (uint8_t i = 0; i < currentSize; i++) {
 			if (value == this->value[i] ) {
 				return this->key[i];
 			}
